+++ HISTORY.rst
@@ -37,6 +37,7 @@
 - Add tag attach and tag remove events.
 - Change the permission needed to attach and remove tags.
 - Add HTTP POST workflow state action.
+- Add access control grant workflow state action.
 
 2.6.4 (2017-07-26)
 ==================